Revert "refactor(worker): move asset route handlers to routes/document-assets.ts"
This reverts commit 892ecc5dfd082a8e9db570fea3a849ff10c59c93.
